Bound Icon View Rubber Banding
This bounds/clamps the rubber banding selection to the actually visible
portion of the AbstractScrollableWidget, and adds a utility function to
the LibGfx Rect that clamps a Point.

With this change the selection rubber band will not go outside the
visible area, and instead stop at the Border of the visible area.